validation, just allow to update and delete post if is the user or admin

+++ Core/Handlers/PostHandler.cs
@@ -2,7 +2,14 @@
 using Blog.Core.Models;
 using Blog.Data;
 using Blog.Data.Entities;
+using Core.Models;
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Identity;
+using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Options;
+using System.Net;
+using System.Security.Claims;
 
 namespace Core.Handlers
 {
@@ -10,10 +17,17 @@ namespace Core.Handlers
     {
         private readonly AppDbContext _context;
         private readonly IMapper _mapper;
-        public PostHandler(AppDbContext context, IMapper mapper)
+        private readonly SignInManager<IdentityUser> _signInManager;
+        private readonly UserManager<IdentityUser> _userManager;
+        private readonly JwtSettings _jwtSettings;
+
+        public PostHandler(AppDbContext context, IMapper mapper, SignInManager<IdentityUser> signInManager, UserManager<IdentityUser> userManager, IOptions<JwtSettings> jwtSettings)
         {
             _context = context;
             _mapper = mapper;
+            _signInManager = signInManager;
+            _userManager = userManager;
+            _jwtSettings = jwtSettings.Value;
         }
 
         public async Task<IEnumerable<PostModel>> GetAll(bool isFromApi)
@@ -29,14 +43,17 @@ namespace Core.Handlers
             return models;
         }
 
-        public async Task<PostModel> Add(PostModel model)
+        public async Task<PostInsertModel> Add(PostInsertModel model, string loggedUser)
         {
+            var identityUser = await _userManager.FindByEmailAsync(loggedUser);
+            var user = await _context.Users.FirstOrDefaultAsync(a => a.IdentityUserId == identityUser.Id);
+
             var entity = _mapper.Map<Post>(model);
             entity.CreatedAt = DateTime.Now;
+            entity.AutorId = user.Id;
+
             _context.Add(entity);
             await _context.SaveChangesAsync();
-
-            model.Id = entity.Id;
 
             return model;
         }
@@ -49,13 +66,24 @@ namespace Core.Handlers
             return _mapper.Map<PostModel>(entity);
         }
 
-        public async Task<PostModel> Edit(PostModel model)
+        public async Task<PostInsertModel> Edit(int id, [FromBody] PostInsertModel model, string loggedUser)
         {
-            var entity = _mapper.Map<Post>(model);
-            _context.Update(entity);
-            await _context.SaveChangesAsync();
+            var entity = await _context.Posts.FindAsync(id);
+            if (entity == null)
+                return null;
 
-            return model;
+            bool allowed = await IsAllowed(loggedUser, entity);
+            if (allowed)
+            {
+                entity.Title = model.Title;
+                entity.Content = model.Content;
+                _context.Update(entity);
+                await _context.SaveChangesAsync();
+
+                return model;
+            }
+
+            return null;
         }
 
         public bool Exists(int id)
@@ -63,11 +91,29 @@ namespace Core.Handlers
             return _context.Posts.Any(e => e.Id == id);
         }
 
-        public async Task Delete(int id)
+        public async Task<int> Delete(int id, string loggedUser)
         {
             var entity = await _context.Posts.FindAsync(id);
-            _context.Posts.Remove(entity);
-            await _context.SaveChangesAsync();
+            bool allowed = await IsAllowed(loggedUser, entity);
+
+            if (allowed)
+            {
+                _context.Posts.Remove(entity);
+                await _context.SaveChangesAsync();
+                return StatusCodes.Status204NoContent;
+            }
+
+            return StatusCodes.Status403Forbidden;
+        }
+
+        private async Task<bool> IsAllowed(string loggedUser, Post entity)
+        {
+            var identityUser = await _userManager.FindByEmailAsync(loggedUser);
+            var roles = await _userManager.GetRolesAsync(identityUser);
+            var user = await _context.Users.FirstOrDefaultAsync(a => a.IdentityUserId == identityUser.Id);
+            bool allowed = _signInManager.Context.User.IsInRole("admin") || entity.AutorId == user.Id;
+
+            return allowed;
         }
     }
 }
